requirements.txt
aptly-api-client==0.1.5
certifi==2017.11.5
chardet==3.0.4
idna==2.6
iso8601==0.1.12
pkg-resources==0.0.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
typing==3.6.2
urllib3==1.22
